@@ -1,15 +1,15 @@
 library chatverse;
 
-export 'src/models/message.dart';
-export 'src/models/chat_user.dart';
-export 'src/models/chat_room.dart';
-export 'src/widgets/chat_view.dart';
-export 'src/widgets/chat_input.dart';
-export 'src/widgets/chat_bubble.dart';
-export 'src/services/chat_service.dart';
-export 'src/services/auth_service.dart';
-export 'src/utils/chat_theme.dart';
 export 'src/chat_controller.dart';
+export 'src/models/chat_room.dart';
+export 'src/models/chat_user.dart';
+export 'src/models/message.dart';
+export 'src/services/chat_service.dart';
+export 'src/widgets/chat_input.dart';
+export 'src/widgets/chat_list_view.dart';
+export 'src/widgets/chat_view.dart';
+export 'src/widgets/chat_room_tile.dart';
+export 'src/utils/chat_theme.dart';
 
 /// A Calculator.
 class Calculator {
